[common] refactoring: 'get_human_size_str' moved from 'system' to 'view_utils'
