+++ bauh/commons/system.py
@@ -22,9 +22,6 @@ if GLOBAL_PY_LIBS not in PATH:
 
 USE_GLOBAL_INTERPRETER = bool(os.getenv('VIRTUAL_ENV'))
 
-SIZE_UNITS = ((1, 'B'), (1024, 'Kb'), (1048576, 'Mb'), (1073741824, 'Gb'),
-              (1099511627776, 'Tb'), (1125899906842624, 'Pb'))
-
 
 def gen_env(global_interpreter: bool, lang: str = DEFAULT_LANG, extra_paths: Optional[Set[str]] = None) -> dict:
     custom_env = dict(os.environ)
@@ -312,22 +309,6 @@ def get_dir_size(start_path='.'):
                 total_size += os.path.getsize(fp)
 
     return total_size
-
-
-def get_human_size_str(size) -> Optional[str]:
-    if type(size) in (int, float, str):
-        int_size = abs(int(size))
-
-        if int_size == 0:
-            return '0'
-
-        for div, unit in SIZE_UNITS:
-
-            size_unit = int_size / div
-
-            if size_unit < 1024:
-                size_unit = size_unit if size > 0 else size_unit * -1
-                return f'{int(size_unit)} {unit}' if unit == 'B' else f'{size_unit:.2f} {unit}'
 
 
 def run(cmd: List[str], success_code: int = 0, custom_user: Optional[str] = None) -> Tuple[bool, str]:
